feat: f64→f32 migration, data_type/is_setpoint to fields, tag_types table, shared migrations
- Rename f64 to f32 across all apps (tag-registry, mqtt-simulator, caro-hmi)
- Move data_type and is_setpoint from top-level template properties to fields
  entries (field_type TagType and Boolean respectively)
- Add Trends as required Boolean field on tag templates
- Add tag_types lookup table (migrations 007, 008, 009); only f32 and bool remain
- validateTemplate enforces three required fields on tag templates
- Server validates TagType field defaults against tag_types table on save
- Shared migrations: all servers (tag-registry, mqtt-simulator, caro-hmi) now run
  runMigrations() at startup with pg_advisory_lock concurrency protection
- hmi-context TagDef.data_type widened to string for forward-compatibility
- HmiContextProvider loading guard restored
- Propagate all deltas to spec docs; clear TR-001, TR-002, TR-003, HMI-001
- Fix CARO_DB_Spec phantom migration 005 reference
- Update all HMI/MQTT docs to reflect f32-only type system

+++ apps/caro-hmi/server/src/__tests__/mqtt-bridge.test.ts
@@ -25,7 +25,7 @@ function makeTagDef(tag_id: number): TagDef {
   return {
     tag_id,
     tag_path:    `${MODULE_ID}.tag${tag_id}`,
-    data_type:   'f64',
+    data_type:   'f32',
     is_setpoint: false,
     module_id:   MODULE_ID,
     eng_min:     null,
