+++ apps/tag-registry/e2e/tests/fields-panel.spec.js
@@ -19,7 +19,7 @@ test.describe('Fields Panel', () => {
     created.push(tName, sName);
 
     // T: tag template with a numeric field
-    await createTagTemplate(tName, 'f64', false, {
+    await createTagTemplate(tName, 'f32', false, {
       eng_min: { field_type: 'Numeric', default: 0 },
     });
